Fix query 5 so it passes validation.
The original query 5 implementation produced the same output as the
Neo4j implementation running on a SF1 database and processing the
associated substition parameters (query_5_param.txt).  Unfortunately
that implementation didn't pass validation.  The SF1 database was
large enough that there always was at least one post in each of the
top 20 forums.  That is not the case with the validation database.

Break the processing into two passes.  The first one identifies all
the forums and resets the counts.  The second counts the posts.  So
even if there are no post to count, all forums will start life with
zero counts.