A correlated subselect not supported by MemSQL Distributed.
Feature 'Correlated subselect that can not be transformed and does not
match on shard keys' is not supported by MemSQL Distributed.

Break up the original single SQL query into a few smaller ones running
in a try block to eliminate use of problematic feature.

Short query 7 now runs with MemSQL/cluster-in-a-box.